Fixed SSL cert availability issue in guest causing nginx start failure on reload
A previous release moved the nginx configs into the guest, and shifted service starts to run on boot vs using a custom provisioner. The ssl certs still begin on a mount, which was unavailable during the boot sequence, broke the nginx service start since it's configuration referenced ssl files it could not read till later.

This moves all the ssl files into the guest except for the shared CA root so that the CA root may maintain trusted status on the host machine.

As a result of this change, custom .nginx.conf files referencing the old shared storage for ssl certs must be updated. That may be accomplished via the following command:

        find /server/sites -type f -name .nginx.conf -mindepth 2 -maxdepth 2 | xargs perl -pi -e 's#/server/.shared/ssl#/etc/nginx/ssl#g'

Post-update (optional) host machine cleanup may be run:

        rm -vf /server/.shared/ssl/*.c??.pem /server/.shared/ssl/local.key.pem